+++ frontend/src/components/modules/details/detailsTitleBarHandler.js
@@ -1,36 +1,57 @@
 import React from 'react';
 import { Typography } from '@material-ui/core';
-import defaultTitleBarHandler from '../default/defaultTitleBarHandler';
+import { getTitleBarStyle, styles } from '../default/defaultTitleBarHandler';
+import PropTypes from 'prop-types';
+import AdvanaMegaMenuPill from '@dod-advana/advana-platform-ui/dist/megamenu/AdvanaMegaMenuPill';
 
-const DetailsTitleBar = {
-	getTitleBar: (props) => {
-		const { detailsType } = props;
-		return (
-			<div style={{ display: 'flex', marginLeft: '3%' }}>
-				<Typography variant="h1" style={styles.wording} display="inline">
-					{detailsType}
-				</Typography>
-				<Typography variant="h1" style={styles.wording} display="inline">
-					Details
-				</Typography>
+const getTitleBar = (props) => {
+	const { detailsType } = props;
+	return (
+		<div style={{ display: 'flex', marginLeft: '3%' }}>
+			<Typography variant="h1" style={styles.wording} display="inline">
+				{detailsType}
+			</Typography>
+			<Typography variant="h1" style={styles.wording} display="inline">
+				Details
+			</Typography>
+		</div>
+	);
+};
+
+const DetailsTitleBar = (props) => {
+	const { style, children, componentStepNumbers = [], jupiter, detailsType, rawSearchResults, pageDisplayed } = props;
+
+	return (
+		<div
+			style={{ ...styles.container, ...style }}
+			className={componentStepNumbers ? `tutorial-step-${componentStepNumbers['Search Bar']}` : null}
+		>
+			<div
+				style={getTitleBarStyle({
+					rawSearchResults,
+					pageDisplayed,
+				})}
+			>
+				{getTitleBar({
+					detailsType,
+				})}
+				<div style={styles.searchBar}>{children}</div>
+				{!jupiter && <AdvanaMegaMenuPill margin="0 -30px 0 20px" defaultHeader="Applications" />}
 			</div>
-		);
-	},
 
-	getCategoryTabs(props) {
-		return <></>;
-	},
+			<></>
+		</div>
+	);
+};
 
-	getTitleBarStyle(props) {
-		return defaultTitleBarHandler.getTitleBarStyle(props);
-	},
+DetailsTitleBar.propTypes = {
+	style: PropTypes.objectOf(PropTypes.string),
+	children: PropTypes.element,
+	componentStepNumbers: PropTypes.objectOf(PropTypes.number),
+	jupiter: PropTypes.bool,
+	detailsType: PropTypes.string,
+	rawSearchResults: PropTypes.array,
+	pageDisplayed: PropTypes.string,
 };
 
 export default DetailsTitleBar;
-
-const styles = {
-	wording: {
-		color: '#131E43',
-		marginRight: 15,
-	},
-};
