+++ frontend/src/components/modules/eda/edaTitleBarHandler.js
@@ -1,49 +1,77 @@
 import React from 'react';
 import ContractSearchLogo from '../../../images/logos/GAMECHANGER-Contract.png';
-import defaultTitleBarHandler from '../default/defaultTitleBarHandler';
+import { getTitleBarStyle, styles } from '../default/defaultTitleBarHandler';
+import PropTypes from 'prop-types';
+import AdvanaMegaMenuPill from '@dod-advana/advana-platform-ui/dist/megamenu/AdvanaMegaMenuPill';
 
-const EdaTitleBarHandler = {
-	getTitleBar: (props) => {
-		const { componentStepNumbers, onTitleClick, cloneData } = props;
-		return (
-			<div
-				className={`tutorial-step-${componentStepNumbers[`${cloneData.display_name} Title`]}`}
+const getTitleBar = (props) => {
+	const { componentStepNumbers, onTitleClick, cloneData } = props;
+	return (
+		<div
+			className={`tutorial-step-${componentStepNumbers[`${cloneData.display_name} Title`]}`}
+			onClick={onTitleClick}
+		>
+			<img
+				src={ContractSearchLogo}
+				style={styles.title}
 				onClick={onTitleClick}
+				alt="contractSearch"
+				className={
+					componentStepNumbers
+						? `tutorial-step-${componentStepNumbers[`${cloneData.display_name} Title`]}`
+						: null
+				}
+			/>
+		</div>
+	);
+};
+
+const EdaTitleBarHandler = (props) => {
+	const {
+		style,
+		children,
+		onTitleClick,
+		componentStepNumbers = [],
+		jupiter,
+		cloneData,
+		rawSearchResults,
+		pageDisplayed,
+	} = props;
+
+	return (
+		<div
+			style={{ ...styles.container, ...style }}
+			className={componentStepNumbers ? `tutorial-step-${componentStepNumbers['Search Bar']}` : null}
+		>
+			<div
+				style={getTitleBarStyle({
+					rawSearchResults,
+					pageDisplayed,
+				})}
 			>
-				<img
-					src={ContractSearchLogo}
-					style={styles.title}
-					onClick={onTitleClick}
-					alt="contractSearch"
-					className={
-						componentStepNumbers
-							? `tutorial-step-${componentStepNumbers[`${cloneData.display_name} Title`]}`
-							: null
-					}
-				/>
+				{getTitleBar({
+					componentStepNumbers,
+					onTitleClick,
+					cloneData,
+				})}
+				<div style={styles.searchBar}>{children}</div>
+				{!jupiter && <AdvanaMegaMenuPill margin="0 -30px 0 20px" defaultHeader="Applications" />}
 			</div>
-		);
-	},
 
-	getCategoryTabs(props) {
-		return <></>;
-	},
+			<></>
+		</div>
+	);
+};
 
-	getTitleBarStyle(props) {
-		return defaultTitleBarHandler.getTitleBarStyle(props);
-	},
+EdaTitleBarHandler.propTypes = {
+	style: PropTypes.objectOf(PropTypes.string),
+	children: PropTypes.element,
+	onTitleClick: PropTypes.func,
+	componentStepNumbers: PropTypes.objectOf(PropTypes.number),
+	jupiter: PropTypes.bool,
+	cloneData: PropTypes.object,
+	rawSearchResults: PropTypes.array,
+	pageDisplayed: PropTypes.string,
 };
 
 export default EdaTitleBarHandler;
-
-const styles = {
-	wording: {
-		color: '#131E43',
-		marginRight: 15,
-	},
-	title: {
-		margin: '0 25px 0 15px',
-		cursor: 'pointer',
-		height: '50px',
-	},
-};
