+++ frontend/src/containers/GameChangerDetailsPage.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useRef } from 'react';
-import SearchBanner from '../components/searchBar/GCSearchBanner';
+import TitleBar from '../components/searchBar/TitleBar';
 import { trackEvent, trackPageView } from '../components/telemetry/Matomo';
 import GameChangerAPI from '../components/api/gameChanger-service-api';
 import { gcColors } from './GameChangerPage';
@@ -917,11 +917,11 @@ const GameChangerDetailsPage = (props) => {
 
 	return (
 		<div style={{ minHeight: 'calc(100% - 89px)', background: 'white' }}>
-			<SearchBanner
+			<TitleBar
 				detailsType={detailsType}
 				titleBarModule={'details/detailsTitleBarHandler'}
 				rawSearchResults={[]}
-			></SearchBanner>
+			></TitleBar>
 
 			{showEntityContainer && renderEntityContainer()}
